+++ src/main/java/at/hannibal2/skyhanni/api/enoughupdates/EnoughUpdatesManager.kt
@@ -108,7 +108,7 @@ object EnoughUpdatesManager {
         val async = list.mapNotNullAsync { name ->
             try {
                 val internalName = name.removeSuffix(".json")
-                val itemJson = fileSystem.readAllBytesAsJsonElement("items/$name").asJsonObject
+                val itemJson = fileSystem.readJson("items/$name").asJsonObject
                 val item = parseItem(
                     internalName = internalName,
                     json = itemJson,
