+++ src/main/java/at/hannibal2/skyhanni/data/repo/ChatProgressUpdates.kt
@@ -196,6 +196,10 @@ class ChatProgressUpdates private constructor(val category: ChatProgressCategory
             return progress
         }
 
+        fun startBlock(label: String, block: (ChatProgressUpdates) -> Unit): ChatProgressUpdates = start(label).apply {
+            block(this)
+        }
+
         fun toggle() {
             enabled = !enabled
             dirty = true
