+++ src/main/java/at/hannibal2/skyhanni/utils/coroutines/CoroutineConfig.kt
@@ -23,7 +23,7 @@ import kotlin.time.Duration.Companion.seconds
 open class CoroutineConfig(
     val name: String,
     val timeout: Duration = 10.seconds,
-    val withIOContext: Boolean = false,
+    private val withIOContext: Boolean = false,
 ) {
     fun withMutex(mutex: Mutex): MutexedCoroutineConfig = MutexedCoroutineConfig(name, mutex, timeout, withIOContext)
     open fun withIOContext(): CoroutineConfig = CoroutineConfig(name, timeout, withIOContext = true)
